+++ robots/cmd/test-label-analyzer/cmd/filter/root.go
@@ -27,24 +27,11 @@ import (
 var rootCmd = &cobra.Command{
 	Use:   "filter",
 	Short: "Filters the output of commands",
-	Long:  `TODO`, // TODO
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("filter called")
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return fmt.Errorf("requires a subcommand to be used")
 	},
 }
 
 func Command() *cobra.Command {
 	return rootCmd
-}
-
-func init() {
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// rootCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
